Massively mangled tunnel vision.  But hey, it compiles.
git-svn-id: file:///home/apenwarr/alumnit-svn/public/trunk@83 6619c942-59fa-0310-bf43-a07aaa81d9ac

+++ utils/wvlinklist.cc
@@ -19,7 +19,7 @@ WvLink::WvLink(void *_data, WvLink *prev, WvLink *&tail, bool _auto_free)
 }
 
 
-int WvList::count() const
+size_t WvList::count() const
 {
     WvLink *l;
     size_t n = 0;
